Add TablePaginationActions to new component
