+++ src/components/TransactionList/TransactionList.tsx
@@ -9,83 +9,10 @@ import Grid from "@material-ui/core/Grid/Grid";
 import Tooltip from "@material-ui/core/Tooltip/Tooltip";
 import TableSortLabel from "@material-ui/core/TableSortLabel/TableSortLabel";
 import TableFooter from '@material-ui/core/TableFooter';
-import { makeStyles, useTheme, Theme, createStyles } from '@material-ui/core/styles';
 import TablePagination from '@material-ui/core/TablePagination';
-import IconButton from '@material-ui/core/IconButton';
-import FirstPageIcon from '@material-ui/icons/FirstPage';
-import KeyboardArrowLeft from '@material-ui/icons/KeyboardArrowLeft';
-import KeyboardArrowRight from '@material-ui/icons/KeyboardArrowRight';
-import LastPageIcon from '@material-ui/icons/LastPage';
 
+import TablePaginationActions from './components/TablePaginationActions'
 import { stableSort, getSorting } from "../../helpers/helpers";
-
-const useStyles1 = makeStyles((theme: Theme) =>
-    createStyles({
-        root: {
-            flexShrink: 0,
-            color: theme.palette.text.secondary,
-            marginLeft: theme.spacing(2.5),
-        },
-    }),
-);
-
-interface TablePaginationActionsProps {
-    count: number;
-    page: number;
-    rowsPerPage: number;
-    onChangePage: (event: React.MouseEvent<HTMLButtonElement, MouseEvent>, newPage: number) => void;
-}
-
-function TablePaginationActions(props: TablePaginationActionsProps) {
-    const classes = useStyles1();
-    const theme = useTheme();
-    const { count, page, rowsPerPage, onChangePage } = props;
-
-    function handleFirstPageButtonClick(event: React.MouseEvent<HTMLButtonElement, MouseEvent>) {
-        onChangePage(event, 0);
-    }
-
-    function handleBackButtonClick(event: React.MouseEvent<HTMLButtonElement, MouseEvent>) {
-        onChangePage(event, page - 1);
-    }
-
-    function handleNextButtonClick(event: React.MouseEvent<HTMLButtonElement, MouseEvent>) {
-        onChangePage(event, page + 1);
-    }
-
-    function handleLastPageButtonClick(event: React.MouseEvent<HTMLButtonElement, MouseEvent>) {
-        onChangePage(event, Math.max(0, Math.ceil(count / rowsPerPage) - 1));
-    }
-
-    return (
-        <div className={classes.root}>
-            <IconButton
-                onClick={handleFirstPageButtonClick}
-                disabled={page === 0}
-                aria-label="first page"
-            >
-                {theme.direction === 'rtl' ? <LastPageIcon /> : <FirstPageIcon />}
-            </IconButton>
-            <IconButton onClick={handleBackButtonClick} disabled={page === 0} aria-label="previous page">
-                {theme.direction === 'rtl' ? <KeyboardArrowRight /> : <KeyboardArrowLeft />}
-            </IconButton>
-            <IconButton
-                onClick={handleNextButtonClick}
-                disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-                aria-label="next page"
-            >
-                {theme.direction === 'rtl' ? <KeyboardArrowLeft /> : <KeyboardArrowRight />}
-            </IconButton>
-            <IconButton
-                onClick={handleLastPageButtonClick}
-                disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-                aria-label="last page"
-            >
-                {theme.direction === 'rtl' ? <FirstPageIcon /> : <LastPageIcon />}
-            </IconButton>
-        </div>
-    );
-}
 
 const headerCols = [
     { id: "timestamp", numeric: false, disablePadding: true, label: "Timestamp" },
@@ -203,43 +130,47 @@ const TransactionList = (): React.ReactElement => {
         )))
     }
 
-    return (
-        <Grid container spacing={9} className="Container">
-            <Grid item xs={12} lg={12}>
-                <Table>
-                    <TableHead>
-                        <TableRow>
-                            {transactionListHeaderGenerate(headerCols)}
-                            <TableCell />
-                            <TableCell />
-                        </TableRow>
-                    </TableHead>
-                    <TableBody>
-                        {transactionListRowsGenerate(blokchain)}
-                    </TableBody>
-                    <TableFooter>
-                        <TableRow>
-                            <TablePagination
-                                rowsPerPageOptions={[15, 25, 50, 100]}
-                                colSpan={3}
-                                count={blokchain.length}
-                                rowsPerPage={rowsPerPage}
-                                page={page}
-                                SelectProps={{
-                                    inputProps: { 'aria-label': 'rows per page' },
-                                    native: true,
-                                }}
-                                onChangePage={handleChangePage}
-                                onChangeRowsPerPage={handleChangeRowsPerPage}
-                                ActionsComponent={TablePaginationActions}
-                            />
-                        </TableRow>
-                    </TableFooter>
+    const tablePaginationProps = {
+        rowsPerPageOptions:  [15, 25, 50, 100],
+        colSpan:  3,
+        count:  blokchain.length ,
+        rowsPerPage: rowsPerPage ,
+        page :  page ,
+        SelectProps: {
+        inputProps: { 'aria-label': 'rows per page' },
+            native: true,
+         },
+        
+        onChangePage :  handleChangePage ,
+        onChangeRowsPerPage :  handleChangeRowsPerPage ,
+        ActionsComponent :  TablePaginationActions 
+    }
 
-                </Table>
-            </Grid>
+return (
+    <Grid container spacing={9} className="Container">
+        <Grid item xs={12} lg={12}>
+            <Table>
+                <TableHead>
+                    <TableRow>
+                        {transactionListHeaderGenerate(headerCols)}
+                        <TableCell />
+                        <TableCell />
+                    </TableRow>
+                </TableHead>
+                <TableBody>
+                    {transactionListRowsGenerate(blokchain)}
+                </TableBody>
+                <TableFooter>
+                    <TableRow>
+                        <TablePagination
+                            {...tablePaginationProps}
+                        />
+                    </TableRow>
+                </TableFooter>
+            </Table>
         </Grid>
-    );
+    </Grid>
+);
 };
 
 export default TransactionList;
